requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.8.0
